Fix issue: speed migrate is broken
To debug this, set_timeout() added to wrap all timeout setting in SPEED thread
Issue was that CHECK_BATTERIES message was overwriting the migrate timer
TO fix it, use the timeout to check the batteries, but only when the motor is
running. Stop sending CHECK BATTERIES, the timeout is sufficient.